requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
dj-database-url>=2.1
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.6
Pillow>=10.0
